Clean root
